[notification hubs] Update APNS messages
