+++ sdk/notificationhubs/notification-hubs/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.2",
   "description": "Azure Notification Hubs SDK for JavaScript",
   "sdk-type": "client",
-  "main": "dist/index.js",
+  "main": "dist/index.cjs",
   "module": "dist-esm/src/index.js",
   "browser": {
     "./dist-esm/src/auth/hmacSha256.js": "./dist-esm/src/auth/hmacSha256.browser.js"
@@ -46,12 +46,11 @@
   "scripts": {
     "audit": "node ../../../common/scripts/rush-audit.js && rimraf node_modules package-lock.json && npm i --package-lock-only 2>&1 && npm audit",
     "build:samples": "echo Obsolete",
-    "build:test:browser": "tsc -p . && rollup -c rollup.test.config.js 2>&1",
+    "build:test:browser": "tsc -p . && rollup --bundleConfigAsCjs -c rollup.test.config.cjs 2>&1",
     "build:test:node": "tsc -p . && dev-tool run bundle",
-    "build:test": "tsc -p . && rollup -c rollup.test.config.js 2>&1",
+    "build:test": "tsc -p . && rollup --bundleConfigAsCjs -c rollup.test.config.cjs 2>&1",
     "build:types": "downlevel-dts types/latest/ types/3.1/",
-    "build:output": "node scripts/renameOutput.mjs",
-    "build": "npm run build:test && api-extractor run --local && npm run build:types && npm run build:output",
+    "build": "npm run build:test && api-extractor run --local && npm run build:types",
     "check-format": "prettier --list-different --config ../../../.prettierrc.json --ignore-path ../../../.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"*.{js,json}\"",
     "clean": "rimraf dist dist-* temp types *.tgz *.log",
     "execute:samples": "dev-tool samples run samples-dev",
@@ -102,18 +101,17 @@
     "@azure-tools/test-recorder": "^3.0.0",
     "@azure/test-utils": "^1.0.0",
     "@microsoft/api-extractor": "^7.31.1",
-    "@rollup/plugin-commonjs": "^24.0.0",
-    "@rollup/plugin-inject": "^5.0.0",
+    "@rollup/plugin-commonjs": "^25.0.2",
+    "@rollup/plugin-inject": "^5.0.3",
     "@rollup/plugin-json": "^6.0.0",
     "@rollup/plugin-multi-entry": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^13.1.3",
-    "@rollup/plugin-replace": "^5.0.0",
-    "@rollup/plugin-typescript": "^11.0.0",
+    "@rollup/plugin-node-resolve": "^15.1.0",
+    "@rollup/plugin-replace": "^5.0.2",
+    "@rollup/plugin-typescript": "^11.1.1",
     "@types/chai": "^4.3.1",
     "@types/mocha": "^7.0.2",
     "@types/node": "^14.0.0",
     "chai": "^4.3.6",
-    "cross-env": "^7.0.3",
     "dotenv": "^16.0.0",
     "downlevel-dts": "^0.10.0",
     "eslint": "^8.16.0",
@@ -135,10 +133,9 @@
     "prettier": "^2.5.1",
     "puppeteer": "^19.2.2",
     "rimraf": "^3.0.2",
-    "rollup": "^2.0.0",
+    "rollup": "^3.25.3",
     "rollup-plugin-shim": "^1.0.0",
-    "rollup-plugin-sourcemaps": "^0.6.3",
-    "ts-node": "^10.0.0",
+    "ts-node": "^10.9.1",
     "typescript": "~5.0.0",
     "util": "^0.12.4"
   },
