+++ sdk/notificationhubs/notification-hubs/src/auth/hmacSha256.browser.ts
@@ -1,7 +1,40 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-/// <reference lib="dom" />
+declare global {
+  class TextEncoder {
+    encode(input?: string): Uint8Array;
+  }
+
+  interface HmacImportParams {
+    name: string;
+    hash: { name: string };
+  }
+
+  interface CryptoKey {
+    algorithm: HmacImportParams;
+    type: string;
+    extractable: boolean;
+    usages: string[];
+  }
+
+  function btoa(input: string): string;
+}
+
+declare const globalThis: {
+  crypto: {
+    subtle: {
+      importKey(
+        format: string,
+        keyData: Uint8Array,
+        algorithm: HmacImportParams,
+        extractable: boolean,
+        usages: string[]
+      ): Promise<CryptoKey>;
+      sign(algorithm: HmacImportParams, key: CryptoKey, data: Uint8Array): Promise<ArrayBuffer>;
+    };
+  };
+};
 
 export async function signString(key: string, toSign: string): Promise<string> {
   const enc = new TextEncoder();
